Modify stuff in replicaServer to ake i work with one process crash

+++ NATE-REPLICA/LibraryWebService/src/replicaLayer/ReplicaServer.java
@@ -73,11 +73,6 @@ public class ReplicaServer extends Thread{
     	//DatagramSocket aSocket = null;
     	MulticastSocket aSocket = null;
     	
-    	if(supportHighAvailability)
-    	{
-        	listener.start();
-        	client.start();
-    	}
     	
 		try{
 			// create socket at agreed port
@@ -213,7 +208,7 @@ public class ReplicaServer extends Thread{
 		if(isStartSFOperation)
 		{
 			supportHighAvailability = false;
-			return startServers();
+			return startServers(false);
 		}
 		else if(isRestartSFOperation)
 		{
@@ -223,7 +218,9 @@ public class ReplicaServer extends Thread{
 		else if(isStartHAOperation)
 		{
 			supportHighAvailability = true;
-			return startServers();
+        	listener.start();
+        	client.start();
+			return startServers(false);
 		}
 		else if(isRestartHAOperation)
 		{
@@ -286,7 +283,7 @@ public class ReplicaServer extends Thread{
 		}
 	}
 	
-	private String startServers()
+	private String startServers(boolean isRestart)
 	{
 		//TODO: modify LibraryServerImpl type
 		serversMap.put("concordia", new LibraryServerImplBase("localhost", 4441, "concordia"));
@@ -297,20 +294,27 @@ public class ReplicaServer extends Thread{
 		holdbackQueue.clear();
 		deliveryQueue.clear();
 
-		//Delete existing log file
-		File file  = new File(logFileName);
-		file.delete();
+		if(!isRestart)
+		{
+			//Delete existing log file
+			File file  = new File(logFileName);
+			file.delete();
+		}
 		
 		return "Replica " + replicaName + " started its servers";
 	}
 	
 	private String restartServers()
 	{
-		startServers();
+		startServers(true);
 		messageSequenceNumber = 1;
 		holdbackQueue.clear();
 		deliveryQueue.clear();
 		updateServers();
+		listener = new HeartbeatListener(replicaName,replicaPort);
+		client = new HeartbeatClient(replicaName);
+    	listener.start();
+    	client.start();
 		numOperationBeforeCrash = 100;
 		return "Replica " + replicaName + " restarted its servers";
 	}
